+++ models.py
@@ -2,12 +2,14 @@ from datetime import datetime
 from enum import Enum
 from random import choice
 from types import DynamicClassAttribute
+from typing import Any, Dict
 from uuid import uuid1, uuid4
 
 from click import echo
 from langchain.chains import ConversationChain
 from langchain.chat_models import ChatOpenAI
 from langchain.memory import ConversationBufferMemory
+from langchain.schema import HumanMessage, SystemMessage
 from sqlalchemy import create_engine, Column, String, DateTime, Text, JSON
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.engine import URL
@@ -77,6 +79,39 @@ class ChatLogs(Base):
     def __str__(self) -> str:
         role = Role[self.role]
         return f"{role.pretty} {str(self.response).strip()}"
+
+class JudgeMemory(ConversationBufferMemory):
+    ai_prefix: str = "AI"
+    human_prefix: str = "Human"
+    system_prefix: str = "System"
+
+    def set_context(self, initial_prompt: str) -> None:
+        """Seed messages for chat"""
+        self.chat_memory.add_message(SystemMessage(content=initial_prompt))
+
+    def save_context(self, inputs: Dict[str, Any], outputs: Dict[str, str]) -> None:
+        """Override: Do not add new responses, only carry forward system message."""
+        pass
+
+class GuesserMemory(ConversationBufferMemory):
+    ai_prefix: str = "AI"
+    human_prefix: str = "Human"
+    system_prefix: str = "System"
+
+    def set_context(self, rules: str, initial_prompt: str) -> None:
+        """Seed messages for chat"""
+        self.chat_memory.add_message(SystemMessage(content=rules))
+        self.chat_memory.add_message(HumanMessage(content=initial_prompt))
+        # empty message because so we don't lose the rules
+        self.chat_memory.add_message(HumanMessage(content=""))
+
+    def save_context(self, inputs: Dict[str, Any], outputs: Dict[str, str]) -> None:
+        """Override: Do not add new responses, only carry forward system message."""
+        _, output_str = self._get_input_output(inputs, outputs)
+        # remove last message
+        self.chat_memory.messages.pop()
+        self.chat_memory.add_ai_message(output_str)
+
 
 class Agent:
     def __init__(
